Google data added to Firebase users collection
