+++ src/containers/index.js
@@ -1,5 +1,5 @@
 import Home from './Home';
-import LogIn from './LogIn';
+import LogIn from './Login';
 import SignUp from './SignUp';
 
 export {LogIn, SignUp, Home};
